chore: relocate component and page
