+++ src/pages/FAQPage.tsx
@@ -2,7 +2,7 @@
 import 'aos/dist/aos.css';
 import AOS from 'aos';
 import {useEffect} from "react";
-import Navbar from '@/components/registrationForm/Navbar';
+import Navbar from '@/components/homePage/Navbar';
 import FAQPageComponent from '@/components/homePage/FAQPageComponent';
 export const FAQPage = () => {
     useEffect(() => {
