+++ src/pages/FrontPage.tsx
@@ -3,8 +3,8 @@ import { Link } from 'react-router-dom';
 import 'aos/dist/aos.css';
 import AOS from 'aos';
 import React,{useEffect} from "react";
-import Navbar from '@/components/registrationForm/Navbar';
-import LandingSection from '@/components/frontPage/LandingSection';
+import Navbar from '@/components/homePage/Navbar';
+import LandingSection from '@/components/homePage/LandingSection';
 import FAQSection from '@/components/homePage/FAQSection';
 import ContactSection from '@/components/homePage/ContactSection';
 import LearningSection from '@/components/homePage/LearningSection';
